blackjack projact

+++ unit_10/calc.py
@@ -24,7 +24,7 @@ operations = {
 
 def calculation():
 
-    num1 = int(input("What is the first number  ? : "))
+    num1 = float(input("What is the first number  ? : "))
 
     for symbol in operations:
         print(symbol)
@@ -33,7 +33,7 @@ def calculation():
 
     while should_continue:
         operation_symbol = input("Pick an operation from the line above : ")
-        num2 = int(input("What is the next number  ? : "))
+        num2 = float(input("What is the next number  ? : "))
         calculation_function = operations[operation_symbol]
         answer = calculation_function(num1, num2)
